ai_test/backend/requirements.txt
allure-pytest
allure-python-commons
Faker
jmespath
PyMySQL
pytest
PyYAML
requests
requests-toolbelt
rsa

python-dotenv
openai
langchain
langchain_core
langchain-openai
langchain_community
langgraph_supervisor
faiss-cpu
langgraph
langsmith

weaviate-client
langchain-weaviate
lightrag-hku[api]
raganything[all]

fastapi
pydantic[email]
uvicorn
tortoise-orm[asyncmy]
bcrypt
python-multipart
gunicorn
PyJWT